pyproject.toml
[project]
name = "personal-health-data-assistant"
version = "0.1.0"
description = "AI-powered personal health data logging and analytics platform"
authors = [{name = "Your Name", email = "your.email@example.com"}]
readme = "README.md"
requires-python = ">=3.11"

dependencies = [
    # Core shared dependencies
    "python-dotenv>=1.0.0",
    "pydantic>=2.0.0",
    "sqlalchemy>=2.0.0",
    "alembic>=1.12.0",
    "psycopg2-binary>=2.9.0",
    "asyncpg>=0.29.0",
    "pytz>=2024.1",
]

[dependency-groups]
ai_data_logger = [
    "langchain>=0.1.0",
    "langgraph>=0.0.40",
    "openai>=1.0.0",
    "arize-phoenix>=3.0.0",
    "requests>=2.31.0",
]

data_workflows = [
    "pandas>=2.0.0",
    "schedule>=1.2.0",
    "httpx>=0.24.0",
]

analytics_workflows = [
    "pandas>=2.0.0",
    "plotly>=5.15.0",
    "scikit-learn>=1.3.0",
]

ai_analytics_assistant = [
    "langchain>=0.1.0",
    "langgraph>=0.0.40",
    "pandas>=2.0.0",
    "matplotlib>=3.7.0",
]

ui = [
    "streamlit>=1.28.0",
    "plotly>=5.15.0",
]

dev = [
    "pytest>=7.4.0",
    "pytest-asyncio>=0.21.0",
    "pytest-cov>=4.1.0",
    "ruff>=0.1.0",
    "black>=23.0.0",
    "mypy>=1.5.0",
    "pre-commit>=3.4.0",
    "types-pytz>=2024.1.0",
]

[tool.ruff]
line-length = 88
target-version = "py311"

[tool.ruff.lint]
select = ["E", "F", "W", "C90", "I", "N", "UP", "B", "A", "S", "T20", "Q"]
ignore = ["E501", "S101"]  # Line too long (handled by black), Use of assert allowed in tests

[tool.black]
line-length = 88
target-version = ['py311']

[tool.mypy]
python_version = "3.11"
warn_return_any = true
warn_unused_configs = true
disallow_untyped_defs = true

[tool.pytest.ini_options]
testpaths = ["tests"]
pythonpath = ["."]
python_files = "test_*.py"
python_functions = "test_*"
addopts = ["-v", "--tb=short", "--strict-markers", "--import-mode=importlib"]
markers = [
    "integration: marks tests as integration tests",
    "unit: marks tests as unit tests",
]